Including shadows_len on common logic.

+++ rust/screen-sim-core/src/pixels_shadow.rs
@@ -1,8 +1,8 @@
 pub const TEXTURE_SIZE: usize = 510;
 
-pub const SHADOWS_LEN: usize = 24;
+pub const SHADOWS_LEN: isize = 24;
 
-pub fn get_shadows() -> [Box<Fn(usize, usize) -> f64>; SHADOWS_LEN] {
+pub fn get_shadows() -> [Box<Fn(usize, usize) -> f64>; SHADOWS_LEN as usize] {
     [
         Box::new(|_i, _j| 255.0),
         Box::new(|i, j| calc_with_log(i, 0) * calc_with_log(j, 0) * 1.0 * 255.0),
